Hack: Provide boosted current when between FET temperature limits
We want to run at higher current when the temperature is in a safe
range. Modify the FET temperature code so that l_temp_fet_start and end
define a range within which the current will be scaled from a boosted
value to the standard maximum. The degree of boost is encoded in
l_temp_motor_start - the configuration variable for the start of the
temperature range for current limitation. This is an absolute number
(ie, if max current is 300, and if the desired maximum value at low
temperature is 480, l_temp_motor_start should be set to 480). Outside
the range encoded by l_temp_fet_start and l_temp_fet_end, the current
will be at the normal maximum value.